requirements.txt
selenium==4.18.1
python-dotenv==1.0.1
schedule==1.2.1
webdriver-manager==4.0.1 